Added 'Pop Culture Survival Guide' books by Quirk.
An interesting little bundle which includes the eBook version of
'Pride and Prejudice and Zombies' (which I already have an a
pBook). The originals are all only available in PDF and Epub formats,
so I will have to convert the ones I want to read on my Kindle using
Calibre. I don't think I'll do the more graphics heavy ones, but the
text ones should work okay.