Completed humble bundle additions but then changed file names.
I have added all the eBooks that I have purchased from Humble
Bundle. There are some duplicates in the originals, and  I have, by
and large, chosen to de-duplicate the entries in the bibliographic
databases.

These entries were originally placed in the kindle.bib file, but this
seemed to be too generic for the files themselves (many of them are
comic cbz files, which don't play nicely on my Kindle). So I have
moved these entries to a humble.bib file, and created a new file
'amazon.bib' which I intend to use to hold entries for Kindle books
purchased via Amazon.